requirements.txt
pandas==1.4.2
seaborn==0.11.2   
matplotlib==3.5.1 
numpy==1.22.0  
regex==2021.11.10
beautifulsoup4==4.10.0 
requests==2.26.0 